create middleware.go for all middleware
